+++ core/hw/maple/maple_devs.h
@@ -167,7 +167,7 @@ extern u8 EEPROM[0x100];
 void load_naomi_eeprom();
 
 // Mouse position and buttons
-extern u32 mo_buttons[4];
+extern u8 mo_buttons[4];
 extern s32 mo_x_abs[4];
 extern s32 mo_y_abs[4];
 extern f32 mo_x_delta[4];
